contact_extractor/requirements.txt
imaplib2>=3.5
pyYAML>=6.0
python-dateutil>=2.8.2

